Move binary sensor to subdir
