+++ components/hs2xx3a/binary_sensor/binary_sensor.py
