+++ components/hs2xx3a/binary_sensor/hs2xx3a_binary_sensor.h
